+++ R/greml.R
@@ -72,7 +72,14 @@ greml <- function(patterns, x, ignore.case = TRUE, vals = FALSE) {
     function(z) grepl(z, x, ignore.case = ignore.case)
   )
 
-  x <- apply(as.matrix(log_mat), 2, any)
-  if (vals) x <- names(x)[which(x)]
-  return(x)
+  log_mat <- as.matrix(log_mat)
+
+  if (dim(log_mat)[2] == 1) {
+    log_mat <- rbind(t(log_mat), t(log_mat))
+  }
+
+  xmat <- apply(log_mat, 2, any)
+
+  if (vals) xmat <- patterns[which(xmat)]
+  return(xmat)
 }